+++ src/main/java/uk/gov/hmcts/reform/professionalapi/configuration/FeignConfiguration.java
@@ -1,12 +1,7 @@
 package uk.gov.hmcts.reform.professionalapi.configuration;
 
-import feign.RequestInterceptor;
 import feign.codec.Encoder;
 import feign.form.spring.SpringFormEncoder;
-
-import java.util.Enumeration;
-import javax.servlet.http.Cookie;
-import javax.servlet.http.HttpServletRequest;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.ObjectFactory;
@@ -15,8 +10,6 @@ import org.springframework.cloud.openfeign.support.SpringEncoder;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Primary;
-import org.springframework.web.context.request.RequestContextHolder;
-import org.springframework.web.context.request.ServletRequestAttributes;
 
 @Configuration
 @Slf4j
@@ -28,34 +21,5 @@ public class FeignConfiguration {
         ObjectFactory<HttpMessageConverters> messageConverters
     ) {
         return new SpringFormEncoder(new SpringEncoder(messageConverters));
-    }
-
-    public RequestInterceptor requestInterceptor() {
-        return requestTemplate -> {
-            ServletRequestAttributes attrs = (ServletRequestAttributes) RequestContextHolder.getRequestAttributes();
-            if (attrs != null) {
-                HttpServletRequest request = attrs.getRequest();
-                // If you take the following method in the cookie
-                // If you take the following method in the cookie
-                Cookie[] cookies = request.getCookies();
-                if (cookies != null && cookies.length > 0) {
-                    for (Cookie cookie : cookies) {
-                        requestTemplate.header(cookie.getName(), cookie.getValue());
-                    }
-                } else {
-                    log.warn("FeignHeadConfiguration", "Failed to get a cookie!");
-                }
-                Enumeration<String> headerNames = request.getHeaderNames();
-                if (headerNames != null) {
-                    while (headerNames.hasMoreElements()) {
-                        String name = headerNames.nextElement();
-                        String value = request.getHeader(name);
-                        requestTemplate.header(name, value);
-                    }
-                } else {
-                    log.warn("FeignHeadConfiguration", "Failed to get request header!");
-                }
-            }
-        };
     }
 }
